Update weatherProgram.py
In the previous version, the data_of_weekdays() method would only return the data of the last day. This happened because the method overwrote the data of the previous, instead of writing to a new dictionary for each day. This version of the function fixes that problem (method starting from line 473). This version also allows for the get_daily_data() method to be called using 2 parameters, instead of one. The new parameter is 'day', and it is used to select the desired day. The 'index' parameter works the same as it did in the previous version. It returns the desired value in the day's data dictionary (based on index).

Also added some comments and overall fixes to methods affected by changes to the methods mentioned earlier. For example, a method that previously requested the daily data from data_of_weekday(), now requests the data of every day, and excludes the undesired day. Similar changes were made to other functions that used to only request the data of one day.