Update weatherProgram.py
Added methods get_daily_icon() and weekday_of_daily_data(). The first is used for getting description icons for each day of the 7-day forecast. The second will be used to determine the order of the weekday labels. Also changed a few other functions to work with the newly added ones and fixed some old problems in the code.