Update weatherProgram.py
I ran into a problem I should've known was coming. When testing the placing of elements, I exceeded the amount of free API calls. This commit implements a fix to that problem. The fix is working around the IP based api call detection. To do this, I'm using an external module called requests_ip_rotator. This is only necessary for the development of this program, as in actual use, the limit shouldn't be exceeded. However, this fix will also help if it does get exceeded in normal use somehow.