Final version
The ready-to-present version of this program. As a whole, there are still some issues, but because I left the finishing of this program to the last week, I didn't have time to implement everything, or fix everything that I wanted. I definitely learned a lot from this program. Mostly about what not to do, when handling as much data as I'm handling here.